+++ tests/test_data_prepare.py
@@ -28,8 +28,7 @@ def test_default(get_data):
     X_train, X_test = get_data
 
     de = DataPrepare()
-    de = de.fit(X_train)
-    X_train = de.transform(X_train)
+    X_train = de.fit_transform(X_train)
     assert isinstance(X_train, pd.DataFrame)
     assert not X_train.empty
 
@@ -46,14 +45,32 @@ def test_default(get_data):
     assert not c_test.empty
 
 
+def test_default_datasets(get_data):
+    for data_id in [179,1461,31,1471,151,1067,1046,1489,1494]:
+        dataset = fetch_openml(data_id=data_id, as_frame=True)
+        X_train, X_test, y_train, y_test = train_test_split(dataset.data, 
+                                                            dataset.target,
+                                                            test_size=0.2, 
+                                                            random_state=RANDOM_SEED,)
+
+        de = DataPrepare()
+        X_train = de.fit_transform(X_train)
+        assert isinstance(X_train, pd.DataFrame)
+        assert not X_train.empty
+        assert not (X_train.isnull().any().any())
+
+        c_test = de.transform(X_test)
+        assert isinstance(c_test, pd.DataFrame)
+        assert not c_test.empty
+        assert not (c_test.isnull().any().any())
+
 def test_encoders(get_data):
     X_train, X_test = get_data
 
     for cat_encoder_name in cat_encoders_names.keys():
         X_train, X_test = get_data
         de = DataPrepare(cat_encoder_names=[cat_encoder_name,])
-        de = de.fit(X_train)
-        c_train = de.transform(X_train)
+        c_train = de.fit_transform(X_train)
         assert isinstance(c_train, pd.DataFrame)
         assert not c_train.empty
 
@@ -76,3 +93,14 @@ def test_clean_nans():
         assert isinstance(df_clean, pd.DataFrame)
         assert not df_clean.empty
         assert not (df_clean.isnull().any().any())
+
+        df_test = pd.DataFrame([[7, 2, np.nan, 0],
+                            [3, np.nan, np.nan, 1],
+                            [9, np.nan, 7, 5],
+                            [np.nan, 3, 10, 4]],
+            columns=list('ABCD'))
+        df_test_clean = clean_nan_encoder.transform(df_test)
+
+        assert isinstance(df_test_clean, pd.DataFrame)
+        assert not df_test_clean.empty
+        assert not (df_test_clean.isnull().any().any())
